Drops remote curl check.
Curl is required on localhost, not target host. This check could fail the playbook, if curl is missing on remote host (where it is actually not needed). 

Decided not to reimplement this check on ansible controller, as it may be not that easy (if you run ansible controller on MacOS or in containers).