feat: Don't use a proxy when doing Alloy readiness check
Add a variable that allows controlling if we should be using or not the
configured proxy settings for the readiness check of Alloy.

This is required if your host must use a proxy for normal operations
(eg. sending your metrics) and you are binding alloy to localhost (which
is most probably unreachable by the proxy).